pyproject.toml
[project]
name = "qbenchsim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.0",
    "quantum-executor>=0.1.0",
]

[tool.uv.workspace]
members = [
    "packages/sacred-parallelism",
]
